create a custom onClick variant that checks to see if the click originated in this element
this feels a lot nicer than having the no-op message. I'm a little surprised that it's ok to have a failing decoder - I guess it's implicit that they may fail, but kind of interesting that the runtime swallows the message. But good to now, in case it's necessary to apply an additional predicate to events before emitting a message in the future.